Refactor: Extract new mod `bin/src/bin_utils.rs`
