+++ crates/bin/src/bin_util.rs
@@ -1,0 +1,30 @@
+use std::{
+    process::{ExitCode, Termination},
+    time::Duration,
+};
+
+use binstall::errors::BinstallError;
+use log::{error, info};
+
+pub enum MainExit {
+    Success(Duration),
+    Error(BinstallError),
+    Report(miette::Report),
+}
+
+impl Termination for MainExit {
+    fn report(self) -> ExitCode {
+        match self {
+            Self::Success(spent) => {
+                info!("Done in {spent:?}");
+                ExitCode::SUCCESS
+            }
+            Self::Error(err) => err.report(),
+            Self::Report(err) => {
+                error!("Fatal error:");
+                eprintln!("{err:?}");
+                ExitCode::from(16)
+            }
+        }
+    }
+}
