+++ crates/bin/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod args;
+pub mod bin_util;
 pub mod entry;
 pub mod install_path;
 pub mod ui;
