+++ crates/bin/src/main.rs
@@ -1,7 +1,4 @@
-use std::{
-    process::{ExitCode, Termination},
-    time::{Duration, Instant},
-};
+use std::time::Instant;
 
 use binstall::{
     errors::BinstallError,
@@ -10,10 +7,10 @@ use binstall::{
         tasks::AutoAbortJoinHandle,
     },
 };
-use log::{debug, error, info};
+use log::debug;
 use tokio::runtime::Runtime;
 
-use cargo_binstall::*;
+use cargo_binstall::{args, bin_util::MainExit, entry, ui};
 
 #[cfg(feature = "mimalloc")]
 #[global_allocator]
@@ -49,27 +46,4 @@ fn main() -> MainExit {
                 .unwrap_or_else(MainExit::Report)
         })
     })
-}
-
-enum MainExit {
-    Success(Duration),
-    Error(BinstallError),
-    Report(miette::Report),
-}
-
-impl Termination for MainExit {
-    fn report(self) -> ExitCode {
-        match self {
-            Self::Success(spent) => {
-                info!("Done in {spent:?}");
-                ExitCode::SUCCESS
-            }
-            Self::Error(err) => err.report(),
-            Self::Report(err) => {
-                error!("Fatal error:");
-                eprintln!("{err:?}");
-                ExitCode::from(16)
-            }
-        }
-    }
 }
